당첨 결과 계산을 위한 WinResult 클래스 생성, 당첨 결과 계산 메소드 이동, lottoService로 로또 구매, 당첨 결과 계산 기능 통합
