+++ src/main/java/kr/codesquad/Application.java
@@ -2,7 +2,6 @@ package kr.codesquad;
 
 import kr.codesquad.io.Console;
 import kr.codesquad.lotto.Lotto;
-import kr.codesquad.lotto.LottoPurchase;
 import kr.codesquad.lotto.LottoService;
 import kr.codesquad.winLotto.WinCount;
 import kr.codesquad.winLotto.WinLotto;
@@ -16,24 +15,20 @@ public class Application {
 
     private final LottoService lottoService;
 
-    private final LottoPurchase lottoPurchase;
-
     public static final int LOTTO_PRICE = 1000;
     Application(
             Console console,
-            LottoService lottoService,
-            LottoPurchase lottoPurchase
+            LottoService lottoService
     ) {
         this.console = console;
         this.lottoService = lottoService;
-        this.lottoPurchase = lottoPurchase;
     }
 
     public void start() {
         int money = console.inputMoney();
         int lottoAmount = money / LOTTO_PRICE;
         console.printAmount(lottoAmount);
-        List<Lotto> lottos = lottoPurchase.buyLotto(money);
+        List<Lotto> lottos = lottoService.buyLotto(money);
         console.printLottoNum(lottos);
         WinLotto winLotto = new WinLotto(console.inputWinNum(), console.inputBonusNum());
         Map<WinCount, Integer> lottoResult = lottoService.makeLottoResult(lottos, winLotto);
