+++ src/main/java/kr/codesquad/Main.java
@@ -1,13 +1,12 @@
 package kr.codesquad;
 
 import kr.codesquad.io.Console;
-import kr.codesquad.lotto.LottoPurchase;
 import kr.codesquad.lotto.LottoService;
 
 public class Main {
 
     public static void main(String[] args) {
-        Application application = new Application(new Console(), new LottoService(), new LottoPurchase());
+        Application application = new Application(new Console(), new LottoService());
         application.start();
     }
